+++ packages/sources/react/src/index.ts
@@ -39,6 +39,8 @@ export type { VtmnSearchProps } from './components/navigation/VtmnSearch/VtmnSea
 // Overlays
 export { VtmnAlert } from './components/overlays/VtmnAlert';
 export type { VtmnAlertProps } from './components/overlays/VtmnAlert';
+export { VtmnPopover } from './components/overlays/VtmnPopover';
+export type { VtmnPopoverProps } from './components/overlays/VtmnPopover';
 export { VtmnToast } from './components/overlays/VtmnToast';
 export type { VtmnToastProps } from './components/overlays/VtmnToast';
 export { VtmnTooltip } from './components/overlays/VtmnTooltip';
@@ -49,12 +51,16 @@ export { VtmnCheckbox } from './components/selection-controls/VtmnCheckbox';
 export type { VtmnCheckboxProps } from './components/selection-controls/VtmnCheckbox';
 export { VtmnChip } from './components/selection-controls/VtmnChip';
 export type { VtmnChipProps } from './components/selection-controls/VtmnChip';
+export { VtmnQuantity } from './components/selection-controls/VtmnQuantity';
+export type { VtmnQuantityProps } from './components/selection-controls/VtmnQuantity';
 export { VtmnRadioButton } from './components/selection-controls/VtmnRadioButton';
 export type { VtmnRadioButtonProps } from './components/selection-controls/VtmnRadioButton';
 export { VtmnToggle } from './components/selection-controls/VtmnToggle';
 export type { VtmnToggleProps } from './components/selection-controls/VtmnToggle';
 
 // Structure
+export { VtmnAccordion } from './components/structure/VtmnAccordion';
+export type { VtmnAccordionProps } from './components/structure/VtmnAccordion';
 export { VtmnDivider } from './components/structure/VtmnDivider';
 export type { VtmnDividerProps } from './components/structure/VtmnDivider';
 export { VtmnSkeleton } from './components/structure/VtmnSkeleton';