+++ packages/sources/svelte/src/components/overlays/VtmnPopover/test/VtmnPopover.spec.js
@@ -11,7 +11,7 @@ describe('VtmnPopover', () => {
     const { container, getByRole } = render(VtmnPopover, {
       id: 'unit-test-id',
     });
-    expect(getPopover(container)).toHaveAttribute('data-position', 'bottom');
+    expect(getPopover(container)).toHaveAttribute('data-position', 'top');
     expect(getPopover(container)).toHaveAttribute(
       'aria-describedby',
       'unit-test-id',
@@ -50,9 +50,5 @@ describe('VtmnPopover', () => {
   test('Should display description slot', () => {
     const { getByText } = render(VtmnPopover, { id: 'unit-test-id' });
     expect(getByText('unit-test body')).toBeVisible();
-  });
-  test('Should display placeholder', () => {
-    const { getByText } = render(VtmnPopover, { id: 'unit-test-id' });
-    expect(getByText('unit-test placeholder')).toBeVisible();
   });
 });