+++ packages/sources/vue/src/index.ts
@@ -37,6 +37,8 @@ export { default as VtmnTag } from './components/indicators/VtmnTag';
 export { default as VtmnSearch } from './components/navigation/VtmnSearch';
 
 // Overlays
+export { default as VtmnAlert } from './components/overlays/VtmnAlert';
+export { default as VtmnPopover } from './components/overlays/VtmnPopover';
 export { default as VtmnTooltip } from './components/overlays/VtmnTooltip';
 
 // Selection controls
@@ -46,5 +48,6 @@ export { default as VtmnRadioButton } from './components/selection-controls/Vtmn
 export { default as VtmnToggle } from './components/selection-controls/VtmnToggle';
 
 // Structure
+export { default as VtmnAccordion } from './components/structure/VtmnAccordion';
 export { default as VtmnDivider } from './components/structure/VtmnDivider';
 export { default as VtmnSkeleton } from './components/structure/VtmnSkeleton';