+++ packages/sources/css/src/components/overlays/popover/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.14](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-popover@0.5.13...@vtmn/css-popover@0.5.14) (2022-04-04)
+
+**Note:** Version bump only for package @vtmn/css-popover
+
+
+
+
+
 ## [0.5.13](https://github.com/Decathlon/vitamin-web/compare/@vtmn/css-popover@0.5.12...@vtmn/css-popover@0.5.13) (2022-03-11)
 
 **Note:** Version bump only for package @vtmn/css-popover